refine prim_factory
replace prim_factory with SwaggerPrimitive, allow users to add their
type/format handlers.
