+++ pyswagger/primitives/_bool.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import
+
+
+def create_bool(_, v, ctx=None):
+    return bool(v)
+
+
