+++ pyswagger/tests/v2_0/test_prim.py
@@ -20,7 +20,7 @@ class SchemaTestCase(unittest.TestCase):
         t = self.app.resolve('#/definitions/Tag')
         self.assertTrue(isinstance(t, objects.Schema))
 
-        v = t._prim_(dict(id=1, name='Hairy'))
+        v = t._prim_(dict(id=1, name='Hairy'), self.app.prim_factory)
         self.assertTrue(isinstance(v, primitives.Model))
         self.assertEqual(v.id, 1)
         self.assertEqual(v.name, 'Hairy')
@@ -44,7 +44,7 @@ class SchemaTestCase(unittest.TestCase):
                 dict(id=1, name='Hairy'),
                 dict(id=2, name='south'),
             ]
-        ))
+        ), self.app.prim_factory)
         self.assertTrue(isinstance(v, primitives.Model))
         self.assertEqual(v.name, 'Buf')
         self.assertEqual(v.photoUrls[0], 'http://flickr.com')
@@ -68,7 +68,7 @@ class SchemaTestCase(unittest.TestCase):
             skill_id=2,
             location="home",
             skill_name="coding"
-        ))
+        ), self.app.prim_factory)
         self.assertTrue(isinstance(v, primitives.Model))
         self.assertEqual(v.id, 1)
         self.assertEqual(v.skill_id, 2)
@@ -85,7 +85,7 @@ class SchemaTestCase(unittest.TestCase):
             id=1,
             location="office",
             boss_name="not you"
-        ))
+        ), self.app.prim_factory)
         self.assertTrue(isinstance(v, primitives.Model))
         self.assertEqual(v.id, 1)
         self.assertEqual(v.location, "office")
@@ -97,42 +97,42 @@ class SchemaTestCase(unittest.TestCase):
         """
         i = self.app.resolve("#/definitions/int")
 
-        self.assertRaises(errs.ValidationError, i._prim_, 200)
-        self.assertRaises(errs.ValidationError, i._prim_, 99)
+        self.assertRaises(errs.ValidationError, i._prim_, 200, self.app.prim_factory)
+        self.assertRaises(errs.ValidationError, i._prim_, 99, self.app.prim_factory)
 
     def test_array_of_int(self):
         """ test array of integer """
         i = self.app.resolve('#/definitions/array_int')
 
         # pass
-        i._prim_([1, 1, 1, 1, 1])
-        i._prim_([1, 1])
+        i._prim_([1, 1, 1, 1, 1], self.app.prim_factory)
+        i._prim_([1, 1], self.app.prim_factory)
 
         # failed
-        self.assertRaises(errs.ValidationError, i._prim_, [1, 1, 1, 1, 1, 1])
-        self.assertRaises(errs.ValidationError, i._prim_, [1])
+        self.assertRaises(errs.ValidationError, i._prim_, [1, 1, 1, 1, 1, 1], self.app.prim_factory)
+        self.assertRaises(errs.ValidationError, i._prim_, [1], self.app.prim_factory)
 
 
     def test_num_multiple_of(self):
         """ test multipleOf """
         i = self.app.resolve("#/definitions/num_multipleOf")
 
-        self.assertRaises(errs.ValidationError, i._prim_, 4)
-        i._prim_(5) # should raise nothing
+        self.assertRaises(errs.ValidationError, i._prim_, 4, self.app.prim_factory)
+        i._prim_(5, self.app.prim_factory) # should raise nothing
 
     def test_str_enum(self):
         """ test str enum """
         e = self.app.resolve("#/definitions/str_enum")
 
-        self.assertRaises(errs.ValidationError, e._prim_, "yellow")
-        e._prim_("green") # should raise nothing
+        self.assertRaises(errs.ValidationError, e._prim_, "yellow", self.app.prim_factory)
+        e._prim_("green", self.app.prim_factory) # should raise nothing
 
     def test_byte(self):
         """ test byte """
         b = self.app.resolve("#/definitions/byte")
 
-        bv = b._prim_("BBBBB")
-        self.assertEqual(str(bv), "BBBBB")
+        bv = b._prim_("BBBBB", self.app.prim_factory)
+        self.assertEqual(str(bv), "BBBBB", self.app.prim_factory)
         self.assertEqual(bv.to_json(), six.b("QkJCQkI="))
 
     def test_date(self):
@@ -140,12 +140,12 @@ class SchemaTestCase(unittest.TestCase):
         d = self.app.resolve("#/definitions/date")
 
         # test input of constructor
-        self.assertEqual(str(d._prim_(float(0))), "1970-01-01")
-        self.assertEqual(str(d._prim_(datetime.date.fromtimestamp(0))), "1970-01-01")
-        self.assertEqual(str(d._prim_(datetime.date.fromtimestamp(0).isoformat())), "1970-01-01")
+        self.assertEqual(str(d._prim_(float(0), self.app.prim_factory)), "1970-01-01")
+        self.assertEqual(str(d._prim_(datetime.date.fromtimestamp(0), self.app.prim_factory)), "1970-01-01")
+        self.assertEqual(str(d._prim_(datetime.date.fromtimestamp(0).isoformat(), self.app.prim_factory)), "1970-01-01")
 
         # to_json
-        dv = d._prim_(float(0))
+        dv = d._prim_(float(0), self.app.prim_factory)
         self.assertEqual(dv.to_json(), "1970-01-01")
 
     def test_date_time(self):
@@ -153,19 +153,19 @@ class SchemaTestCase(unittest.TestCase):
         d = self.app.resolve("#/definitions/date-time")
 
         # test input of constructor
-        self.assertEqual(str(d._prim_(float(0))), "1970-01-01T00:00:00")
-        self.assertEqual(str(d._prim_(datetime.datetime.utcfromtimestamp(0))), "1970-01-01T00:00:00")
-        self.assertEqual(str(d._prim_(datetime.datetime.utcfromtimestamp(0).isoformat())), "1970-01-01T00:00:00")
+        self.assertEqual(str(d._prim_(float(0), self.app.prim_factory)), "1970-01-01T00:00:00")
+        self.assertEqual(str(d._prim_(datetime.datetime.utcfromtimestamp(0), self.app.prim_factory)), "1970-01-01T00:00:00")
+        self.assertEqual(str(d._prim_(datetime.datetime.utcfromtimestamp(0).isoformat(), self.app.prim_factory)), "1970-01-01T00:00:00")
 
         # to_json
-        dv = d._prim_(float(0))
+        dv = d._prim_(float(0), self.app.prim_factory)
         self.assertEqual(dv.to_json(), "1970-01-01T00:00:00")
 
     def test_model_bool(self):
         """ test a model containing boolean """
         d = self.app.resolve("#/definitions/model_bool")
 
-        dv = d._prim_(dict(bool_val=True)) 
+        dv = d._prim_(dict(bool_val=True), self.app.prim_factory) 
         # try to access it
         self.assertEqual(dv.bool_val, True)
 
@@ -181,7 +181,7 @@ class HeaderTestCase(unittest.TestCase):
         p1 = self.app.resolve(jp_compose(['#', 'paths', '/t', 'get', 'parameters', '0']))
         self.assertTrue(isinstance(p1, objects.Parameter))
 
-        v = p1._prim_([1, 2, 3, 4, 5])
+        v = p1._prim_([1, 2, 3, 4, 5], self.app.prim_factory)
         self.assertTrue(isinstance(v, primitives.Array))
         self.assertEqual(str(v), '1,2,3,4,5')
 
@@ -190,8 +190,8 @@ class HeaderTestCase(unittest.TestCase):
         p2 = self.app.resolve(jp_compose(['#', 'paths', '/t', 'get', 'parameters', '1']))
         self.assertTrue(isinstance(p2, objects.Parameter))
 
-        self.assertRaises(errs.ValidationError, p2._prim_, 101)
-        self.assertRaises(errs.ValidationError, p2._prim_, -1)
+        self.assertRaises(errs.ValidationError, p2._prim_, 101, self.app.prim_factory)
+        self.assertRaises(errs.ValidationError, p2._prim_, -1, self.app.prim_factory)
 
     def test_multi_level_array(self):
         """ header in array of array """
@@ -213,8 +213,12 @@ class HeaderTestCase(unittest.TestCase):
                     [11,12],
                     [13,14]
                 ]
-            ]
-        )), '1|2,3|4,5|6 7|8,9|10 11|12,13|14')
+            ], self.app.prim_factory)), '1|2,3|4,5|6 7|8,9|10 11|12,13|14')
+
+        def test_header_in_response(self):
+            """ header in response """
+            # TODO: we didn't implement converting from
+            # '1,1,1,1' to array yet.
 
 
 class AdditionalPropertiesTestCase(unittest.TestCase):
@@ -225,7 +229,7 @@ class AdditionalPropertiesTestCase(unittest.TestCase):
         kls.app = SwaggerApp._create_(get_test_data_folder(version='2.0', which=os.path.join('schema', 'additionalProperties')))
 
     def test_with_schema(self):
-        m = primitives.prim_factory(
+        m = self.app.prim_factory.produce(
             self.app.resolve('#/definitions/add_prop'),
             dict(
                 name_of_map='test',
@@ -254,7 +258,7 @@ class AdditionalPropertiesTestCase(unittest.TestCase):
 
     def test_with_bool(self):
         d = self.app.resolve('#/definitions/add_prop_bool')
-        m = primitives.prim_factory(
+        m = self.app.prim_factory.produce(
             d,
             dict(
                 name='test_bool',
@@ -270,7 +274,7 @@ class AdditionalPropertiesTestCase(unittest.TestCase):
 
     def test_with_bool_false(self):
         d = self.app.resolve('#/definitions/add_prop_false')
-        m = primitives.prim_factory(
+        m = self.app.prim_factory.produce(
             d,
             dict(
                 name='test_bool',
@@ -290,7 +294,7 @@ class AdditionalPropertiesTestCase(unittest.TestCase):
         additionalProperties intecept all keys
         """
         d = self.app.resolve('#/definitions/add_prop_nested')
-        self.assertRaises(errs.ValidationError, primitives.prim_factory,
+        self.assertRaises(errs.ValidationError, self.app.prim_factory.produce,
             d,
             dict(
                 my_int=99
